refactor(analysis): fold the last private copy of the URL trim into the shared rule
deobfuscate.ts carried a fourth copy of the trailing-punctuation strip, so a URL
named inside a decoded PowerShell payload became a different indicator from the
same URL read by an importer.

It could not simply import the module the importers share: deobfuscate.ts is
analysis/privacy (tier 1) and they are analysis/ingest (tier 3), and an import
may go down a tier or sideways but never up. The rule therefore moves to the
PLATFORM layer, at ingest/textUriTrim.ts beside titleSlug.ts and imageFormat.ts,
which are pure helpers of the same kind. Platform sits below every domain, so all
four scrapers reach it and none keeps a private copy.

Folding this one in also fixes a gap the others never had. deobfuscate's URL_RE
does not exclude "]", so a bracket can genuinely end a match, and its private
class did not strip one — `[http://evil.test/a]` was stored complete with the
bracket. The shared class does strip it. Its class also listed "'", '"' and ">",
none of which URL_RE admits inside a match, so dropping them loses nothing.

Two bookkeeping steps the gates require. The stale flatAnalysisFiles entry is
removed, because that map may not name a file that has left src/analysis/. And
the three new analysis→platform edges are legal downward imports, so
check:boundaries still reports "none new" while the ARCHITECTURE.md comply/total
sentence moves 1,940/1,979 → 1,943/1,982; moduleMap.test.ts asserts those agree,
and the figures come from check-boundaries --json.
