+++ companion/src/analysis/veloTextIocs.ts
@@ -7,7 +7,7 @@
 // the exact thing the rule fired on — is otherwise missed entirely.
 import { addIoc, cleanIp, str, getCI, type SiemIoc } from "./siemImport.js";
 import { extractDomains } from "./textDomains.js";
-import { trimSentencePunctuation } from "./textUriTrim.js";
+import { trimSentencePunctuation } from "../ingest/textUriTrim.js";
 
 const TEXT_URL = /\bhttps?:\/\/[^\s"'<>)\]}]+/gi;
 // Octet-bounded, so a "10.0.22000" version string is not read as an address.
